chore: replacing input
